+++ src/app/globals.css
@@ -6,7 +6,7 @@
   :root {
     --max-content-width: 1352px;
     --main-layout-height: calc(100vh - 8rem - 4rem);
-    --background: 220 13% 91%;
+    --background: 0 0% 100%;
     --foreground: 222.2 84% 4.9%;
     --card: 0 0% 100%;
     --card-foreground: 222.2 84% 4.9%;
@@ -35,6 +35,7 @@
     --box: 0 0% 100%;
     --box-foreground: 215 14% 34%;
 
+    --background-gray: 220 13% 91%;
     --border-and-divide: 240 5% 49%;
   }
 
@@ -63,6 +64,8 @@
     --chart-3: 30 80% 55%;
     --chart-4: 280 65% 60%;
     --chart-5: 340 75% 55%;
+
+    --background-gray: 222.2 84% 4.9%;
 
     --box: 217.2 32.6% 17.5%;
     --box-foreground: 210 40% 98%;
