+++ src/components/write/post-form.tsx
@@ -1,7 +1,7 @@
 import { FormEventHandler } from 'react';
 import Input from '@/components/common/input';
 import Editor from '@/components/write/editor';
-import Button from '@/components/common/button';
+import { Button } from '@/components/ui/button';
 import { TUseInputReturn } from '@/hooks/useInput';
 
 type PostFormProps = {
